EnterArbitrage condition now allows a gap
Example: 0.4% gap between BTCUSDT spot and BTCUSDT Perpetual.
Perpetual pair is on top of spot pair (most of the time) because the funding rate is positive. We can enter this position even if they are not equal, because it is a profit when exiting. The condition to exit is the superposition of both perpetual and spot, and this happens a lot.